+++ elements/test_apps/main/elements/gmf_audio_el_com.h
@@ -95,6 +95,8 @@ void ch_cvt_config_callback(esp_gmf_element_handle_t self, void *ctx);
 
 void rate_cvt_config_callback(esp_gmf_element_handle_t self, void *ctx);
 
+void asrc_config_callback(esp_gmf_element_handle_t self, void *ctx);
+
 void drc_config_callback(esp_gmf_element_handle_t self, void *ctx);
 
 void mbc_config_callback(esp_gmf_element_handle_t self, void *ctx);